# core-banking-cards-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-cards
  description: Banking Cards Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/cards
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/cards/{cardId}/configurations/{configId}:
    get:
      tags:
        - Card Configurations
      summary: Get Card Configuration by ID
      description: Retrieve a specific card configuration record by its unique identifier, ensuring it belongs to the specified card.
      operationId: getConfiguration
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card
          required: true
          schema:
            type: string
            format: uuid
        - name: configId
          in: path
          description: Unique identifier of the card configuration record
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the card configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardConfigurationDTO'
        '404':
          description: Card configuration not found
    put:
      tags:
        - Card Configurations
      summary: Update Card Configuration
      description: Update an existing configuration record associated with the specified card.
      operationId: updateConfiguration
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card
          required: true
          schema:
            type: string
            format: uuid
        - name: configId
          in: path
          description: Unique identifier of the card configuration record to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardConfigurationDTO'
        required: true
      responses:
        '200':
          description: Card configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardConfigurationDTO'
        '404':
          description: Card configuration record not found
    delete:
      tags:
        - Card Configurations
      summary: Delete Card Configuration
      description: Remove an existing card configuration record by its unique identifier.
      operationId: deleteConfiguration
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card
          required: true
          schema:
            type: string
            format: uuid
        - name: configId
          in: path
          description: Unique identifier of the card configuration record to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card configuration record deleted successfully
        '404':
          description: Card configuration record not found
...
  /api/v1/cards/{cardId}:
    get:
      tags:
        - Cards
      summary: Get Card by ID
      description: |-
        Retrieve an existing card record by its unique identifier.

        ## Description
        This endpoint returns comprehensive information about a card, including its status, associated accounts, limits, and key attributes. It provides all the details needed for card management and transaction processing.

        ## Response Details
        The response includes:
        * Basic card information (masked number, expiration date, status)
        * Associated entities (issuer, network, program)
        * Account details (account ID, party ID)
        * Card limits and restrictions
        * Card features and capabilities

        ## Security Considerations
        Sensitive information like the full card number, CVV, and PIN are masked or excluded from the response for security reasons. The card number is partially masked (e.g., '411111******1111') to allow identification while protecting sensitive data.

        ## Common Use Cases
        * Retrieving card details for display in party portals
        * Checking card status before processing transactions
        * Verifying card limits and restrictions
      operationId: getCard
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '404':
          description: Card not found
    put:
      tags:
        - Cards
      summary: Update Card
      description: |-
        Update an existing card record by its unique identifier.

        ## Description
        This endpoint allows modification of card attributes such as status, limits, and expiration details. It enables financial institutions to manage card lifecycle events and adjust card parameters based on party needs or risk assessments.

        ## Updatable Fields
        * `cardStatus` - Current status (ACTIVE, INACTIVE, BLOCKED, etc.)
        * `cardHolderName` - Name displayed on the card
        * `dailyLimit` - Maximum daily transaction amount
        * `monthlyLimit` - Maximum monthly transaction amount
        * `creditLimit` - Credit limit (for credit cards)
        * `expirationDate` - Card expiration date (for extensions)
        * `isActive` - Whether the card is active
        * `isLocked` - Whether the card is temporarily locked

        ## Immutable Fields
        Note that certain sensitive fields and core attributes cannot be modified after card creation:
        * Card number
        * BIN
        * Card type
        * Card network
        * Issuer

        For changes to these immutable attributes, a new card must be issued.

        ## Common Use Cases
        * Activating a newly issued card
        * Blocking a card due to suspected fraud
        * Adjusting card limits based on party request
        * Updating cardholder information
      operationId: updateCard
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDTO'
        required: true
      responses:
        '200':
          description: Card updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '404':
          description: Card not found
    delete:
      tags:
        - Cards
      summary: Delete Card
      description: |-
        Remove an existing card record by its unique identifier.

        ## Description
        This endpoint permanently removes a card from the system. It completely erases the card record and all its associated data from the database.

        ## Warning
        ⚠️ **This operation should be used with extreme caution!**

        Deleting a card has several important implications:
        * All transaction history associated with the card will be orphaned
        * Audit trails will be incomplete
        * Regulatory compliance may be compromised
        * Customer service inquiries about the card cannot be addressed

        ## Recommended Alternative
        In most production scenarios, cards should be **deactivated** or **blocked** rather than deleted. This preserves the card record and its history while preventing any new transactions.

        ## Appropriate Use Cases
        Card deletion should be limited to:
        * Test cards in non-production environments
        * Cards created in error that have never been activated
        * Specific regulatory requirements that mandate complete removal

        ## Required Permissions
        This operation typically requires elevated administrative privileges.
      operationId: deleteCard
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card deleted successfully
        '404':
          description: Card not found
...
  /api/v1/cards/{cardId}/configurations:
    get:
      tags:
        - Card Configurations
      summary: List Card Configurations
      description: Retrieve a paginated list of configuration records associated with the specified card.
      operationId: getAllConfigurations
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Successfully retrieved the card configurations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No card configurations found for the specified card
    post:
      tags:
        - Card Configurations
      summary: Create Card Configuration
      description: Create a new configuration record and associate it with the specified card.
      operationId: createConfiguration
      parameters:
        - name: cardId
          in: path
          description: Unique identifier of the card
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardConfigurationDTO'
        required: true
      responses:
        '201':
          description: Card configuration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardConfigurationDTO'
        '400':
          description: Invalid configuration data provided
...
components:
  schemas:
...
    CardConfigurationDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        cardConfigurationId:
          type: string
          format: uuid
          readOnly: true
        cardId:
          type: string
          format: uuid
        programId:
          type: string
          format: uuid
        configKey:
          type: string
        configValue:
          type: string
        configType:
          type: string
        isActive:
          type: boolean
        activationDate:
          type: string
          format: date-time
        expirationDate:
          type: string
          format: date-time
        isSystemDefault:
          type: boolean
        isProgramDefault:
          type: boolean
        isPartyConfigurable:
          type: boolean
        allowedValues:
          type: string
        minValue:
          type: string
        maxValue:
          type: string
        description:
          type: string
...
    CardDTO:
      required:
        - accountId
        - binId
        - cardHolderId
        - cardHolderName
        - cardNetworkId
        - cardNumber
        - cardStatus
        - cardTypeId
        - currencyCode
        - expirationMonth
        - expirationYear
        - isActive
        - isLocked
        - isPhysical
        - isPrimary
        - isVirtual
        - issuerId
        - partyId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        cardId:
          type: string
          format: uuid
          readOnly: true
        cardNumber:
          type: string
        maskedCardNumber:
          pattern: ^[0-9*]{13,19}$
          type: string
        cardSequenceNumber:
          pattern: ^[0-9]{1,3}$
          type: string
        binId:
          type: string
          format: uuid
        cardTypeId:
          type: string
          format: uuid
        cardNetworkId:
          type: string
          format: uuid
        issuerId:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        partyId:
          type: string
          format: uuid
        cardStatus:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
            - CANCELLED
        cardHolderName:
          maxLength: 100
          minLength: 2
          type: string
        cardHolderId:
          maxLength: 50
          minLength: 5
          type: string
        expirationMonth:
          maximum: 12
          minimum: 1
          type: integer
          format: int32
        expirationYear:
          maximum: 2050
          minimum: 2024
          type: integer
          format: int32
        cvv:
          pattern: ^[0-9]{3,4}$
          type: string
        pin:
          pattern: ^[0-9]{4,8}$
          type: string
        activationDate:
          type: string
          format: date-time
        issuanceDate:
          type: string
          format: date-time
        expirationDate:
          type: string
          format: date-time
        lastUsedDate:
          type: string
          format: date-time
        isPhysical:
          type: boolean
        isVirtual:
          type: boolean
        isPrimary:
          type: boolean
        isActive:
          type: boolean
        isLocked:
          type: boolean
        lockReason:
          maxLength: 255
          minLength: 0
          type: string
        dailyLimit:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: true
          type: number
          format: double
        monthlyLimit:
          maximum: 9999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: true
          type: number
          format: double
        creditLimit:
          maximum: 99999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: true
          type: number
          format: double
        availableBalance:
          minimum: 0
          exclusiveMinimum: false
          type: number
          format: double
        currencyCode:
          pattern: ^[A-Z]{3}$
          type: string
        designId:
          type: string
          format: uuid
        notes:
          maxLength: 1000
          minLength: 0
          type: string
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...